Issue-30: Moved vegalayoutmanager to separate module.
