+++ app/src/main/res/layout/activity_main.xml
@@ -70,7 +70,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="vegaLayoutManager"
-            android:text="@string/vega.layout.manager" />
+            android:text="@string/vega_layout_manager" />
 
         <Button
             android:layout_width="match_parent"
