+++ app/src/main/res/values/strings.xml
@@ -17,8 +17,6 @@
     <string name="duo.navigation.drawer">DuoNavigationDrawer</string>
     <string name="duo.navigation.menu.item">Menu item</string>
 
-    <string name="vega.layout.manager">Vega Layout Manager</string>
-
     <string name="fancy.toast">Fancy Toast</string>
 
     <string name="left">Left</string>
