+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':alerter', ':loadingbutton', ':ticker', ':particle', ':swipeback', ':scalinglayout', ':expandablelayout', ':focusresize'
+include ':app', ':alerter', ':loadingbutton', ':ticker', ':particle', ':swipeback', ':scalinglayout', ':expandablelayout', ':focusresize', ':vegalayoutmanager'
