+++ vegalayoutmanager/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.example.vegalayoutmanager">
+    <application>
+        <activity android:name=".VegaManagerActivity"
+            android:label="@string/vega_layout_manager"/>
+    </application>
+</manifest>
