+++ vegalayoutmanager/src/main/java/com/example/vegalayoutmanager/ImageObject.kt
@@ -1,0 +1,5 @@
+package com.example.vegalayoutmanager
+
+import androidx.annotation.DrawableRes
+
+data class ImageObject(val clubName: String, val country: String, @DrawableRes val clubBackground: Int)
