+++ vegalayoutmanager/src/main/res/values/dimens.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="focus_resize_item_height">100dp</dimen>
+</resources>
